requirements.txt
auth==0.5.3
blinker==1.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
dnspython==1.16.0
eventlet==0.29.1
falcon==2.0.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
requests==2.25.0
requests-jwt==0.5.3
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
Werkzeug==1.0.1
